AddedPersonalDataRequestReadOny

+++ STMApi/Services/IService.cs
@@ -2,10 +2,10 @@
 {
     public interface IService<T> where T : class
     {
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
         Task CreateAsync(T entity);
         Task UpdateAsync(T entity);
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
     }
 }
