+++ STMData/Repositories/IRepository.cs
@@ -8,10 +8,10 @@ namespace STMData.Repositories
 {
     public interface IRepository<T> where T : class
     {   
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
         Task CreateAsync(T entity);
         Task UpdateAsync(T entity);
         Task<IEnumerable<T>> GetAllAsync();
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
     }
 }
